# backend/src/rest/v1/openapi.yaml
openapi: 3.0.3
info:
  title: ChatFlow API
  description: A minimal but clean WhatsApp-style messaging app API
  version: 1.0.0
  contact:
    name: ChatFlow Team
servers:
  - url: http://localhost:3002/v1
    description: Development server
  - url: https://chatflow-backend-3w6u4kmniq-ue.a.run.app/v1
    description: Production server (Cloud Run)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Conversation:
      type: object
      properties:
        id:
          type: string
          pattern: '^conv_[0-9]+_[a-z0-9]+$'
          description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
          example: 'conv_1750386041311_fpmswok2p'
        createdBy:
          type: string
          format: email
        type:
          type: string
          enum: [DIRECT, GROUP]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ConversationParticipant'
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          pattern: '^conv_[0-9]+_[a-z0-9]+$'
          description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
          example: 'conv_1750386041311_fpmswok2p'
        senderId:
          type: string
          format: email
        messageType:
          type: string
          enum: [TEXT, IMAGE, FILE]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ConversationParticipant:
      type: object
      properties:
        userId:
          type: string
          format: email
        joinedAt:
          type: string
          format: date-time
        role:
          type: string
          enum: [ADMIN, MEMBER]
    PaginationResult:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data: {}
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details: {}
...
    CreateConversationRequest:
      type: object
      required:
        - participantEmails
      properties:
        participantEmails:
          type: array
          items:
            type: string
            format: email
          minItems: 1
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details: {}
...
paths:
...
  /conversations:
    get:
      tags:
        - Conversations
      summary: Get user's conversations
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Conversations retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginationResult'
                          - type: object
                            properties:
                              data:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Conversation'

    post:
      tags:
        - Conversations
      summary: Create a new conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Conversation'
...
  /conversations/{conversationId}/messages/{messageId}:
    get:
      tags:
        - Messages
      summary: Get a specific message in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            pattern: '^conv_[0-9]+_[a-z0-9]+$'
            description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
            example: 'conv_1750386041311_fpmswok2p'
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            pattern: '^msg_[0-9]+_[a-z0-9]+$'
            description: 'Message ID in format: msg_{timestamp}_{randomString}'
            example: 'msg_1750386041311_abc123def'
      responses:
        '200':
          description: Message retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Messages
      summary: Update a specific message in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            pattern: '^conv_[0-9]+_[a-z0-9]+$'
            description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
            example: 'conv_1750386041311_fpmswok2p'
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            pattern: '^msg_[0-9]+_[a-z0-9]+$'
            description: 'Message ID in format: msg_{timestamp}_{randomString}'
            example: 'msg_1750386041311_abc123def'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMessageRequest'
      responses:
        '200':
          description: Message updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Messages
      summary: Delete a specific message in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            pattern: '^conv_[0-9]+_[a-z0-9]+$'
            description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
            example: 'conv_1750386041311_fpmswok2p'
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            pattern: '^msg_[0-9]+_[a-z0-9]+$'
            description: 'Message ID in format: msg_{timestamp}_{randomString}'
            example: 'msg_1750386041311_abc123def'
      responses:
        '204':
          description: Message deleted successfully
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /search/conversations:
    get:
      tags:
        - Search
      summary: Search conversations with semantic understanding
      description: Perform intelligent search across conversations using natural language queries
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 500
          description: Search query (supports natural language)
          example: "lunch plans with Sarah"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Maximum number of results to return
      responses:
        '200':
          description: Search results with semantic relevance scoring
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Search service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Search
      summary: Search conversations (POST version for web clients)
      description: Same functionality as GET but via POST to potentially reduce OPTIONS preflight overhead in some scenarios
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - q
              properties:
                q:
                  oneOf:
                    - type: string
                      minLength: 1
                      maxLength: 500
                      description: "Plain text search query (not recommended for production)"
                      example: "lunch plans with Sarah"
                    - $ref: '#/components/schemas/EncryptedField'
                  description: "Search query - should be encrypted for privacy"
                limit:
                  type: integer
                  minimum: 1
                  maximum: 50
                  default: 20
                  description: Maximum number of results to return
      responses:
        '200':
          description: Search results with semantic relevance scoring
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Search service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...